refactor: 패키지명 수정 및 불필요한 어노테이션 제거, UserStatus 생성 방식 변경
- 디렉토리명을 DTO -> dto로 변경하였으나 제대로 인식하지 못해 dto -> dto2로 변경

- DiscodeApplication의 메서드를 변경사항에 맞추어 변경

- 인터페이스, 도메인에 불필요한 @Repository 어노테이션 제거

- BasicUserService에서 userStatus를 생성하는 부분 변경

- UserCreateRequest에서 UserStatus status 제거

- 불필요한 파일 제거
