+++ src/main/java/com/sprint/mission/discodeit/DiscodeitApplication.java
@@ -1,24 +1,25 @@
 package com.sprint.mission.discodeit;
 
-import com.sprint.mission.discodeit.DTO.UserCreateRequest;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.dto2.MessageCreateRequest;
+import com.sprint.mission.discodeit.dto2.PublicChannelCreateRequest;
+import com.sprint.mission.discodeit.dto2.UserCreateRequest;
+import com.sprint.mission.discodeit.entity.UserStatus;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
-import com.sprint.mission.discodeit.service.ServiceFactory;
 import com.sprint.mission.discodeit.service.UserService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.annotation.ComponentScan;
 
+import java.time.Instant;
 import java.util.UUID;
 
 @SpringBootApplication
 public class DiscodeitApplication {
     static UUID setupUser(UserService userService) {
         // UserCreateRequest DTO 생성
-        UserCreateRequest request = new UserCreateRequest("woody", "woody@codeit.com", "woody@codeit.com", null, null);
+        UserCreateRequest request = new UserCreateRequest("woody", "woody@codeit.com", "woody@codeit.com", null);
 
         // 변경된 create 메서드 호출
         UUID userId = userService.create(request);
@@ -28,13 +29,15 @@ public class DiscodeitApplication {
     }
 
     static UUID setupChannel(ChannelService channelService) {
-        UUID channelId = channelService.create(ChannelType.PUBLIC, "공지", "공지 채널입니다.");
+        PublicChannelCreateRequest request = new PublicChannelCreateRequest("공지", "공지사항을 안내하는 채널입니다.");
+        UUID channelId = channelService.createPublicChannel(request);
         System.out.println("채널 생성: " + channelId);
         return channelId;
     }
 
     static void messageCreateTest(MessageService messageService, UUID senderId, UUID channelId) {
-        UUID message = messageService.create("안녕하세요.", senderId, channelId);
+        MessageCreateRequest request = new MessageCreateRequest("안녕하세요",senderId, channelId, null);
+        UUID message = messageService.create(request);
         System.out.println("메시지 생성: " + message);
     }
