+++ src/main/java/com/sprint/mission/discodeit/dto2/ChannelResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto2;
 
 import com.sprint.mission.discodeit.entity.ChannelType;
