+++ src/main/java/com/sprint/mission/discodeit/dto2/MessageUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto2;
 
 import java.util.UUID;
