+++ src/main/java/com/sprint/mission/discodeit/dto2/PrivateChannelCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto2;
 
 import java.util.List;
 import java.util.UUID;
