+++ src/main/java/com/sprint/mission/discodeit/dto2/ReadStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto2;
 
 import java.time.Instant;
 import java.util.UUID;
