+++ src/main/java/com/sprint/mission/discodeit/dto2/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto2;
 
 import com.sprint.mission.discodeit.entity.UserStatus;
 
@@ -6,6 +6,5 @@ public record UserCreateRequest(
         String username,
         String email,
         String password,
-        byte[] profileImage,
-        UserStatus status
+        byte[] profileImage
 ) {}
