+++ src/main/java/com/sprint/mission/discodeit/dto2/UserLoginRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto2;
 
 public record UserLoginRequest(
         String username,
