+++ src/main/java/com/sprint/mission/discodeit/repository/BinaryContentRepository.java
@@ -2,12 +2,10 @@
 package com.sprint.mission.discodeit.repository;
 
 import com.sprint.mission.discodeit.entity.BinaryContent;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.UUID;
 
-@Repository
 public interface BinaryContentRepository {
     // 파일 저장 (Create)
     UUID upsert(BinaryContent binaryContent);
