+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto.UserLoginRequest;
-import com.sprint.mission.discodeit.dto.UserResponse;
+import com.sprint.mission.discodeit.dto2.UserLoginRequest;
+import com.sprint.mission.discodeit.dto2.UserResponse;
 public interface AuthService {
     UserResponse login(UserLoginRequest request);  // 로그인 기능 구현
 }
