+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -1,8 +1,8 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto.UserCreateRequest;
-import com.sprint.mission.discodeit.dto.UserResponse;
-import com.sprint.mission.discodeit.dto.UserUpdateRequest;
+import com.sprint.mission.discodeit.dto2.UserCreateRequest;
+import com.sprint.mission.discodeit.dto2.UserResponse;
+import com.sprint.mission.discodeit.dto2.UserUpdateRequest;
 
 import java.util.List;
 import java.util.UUID;
