+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicAuthService.java
@@ -1,19 +1,17 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.UserLoginRequest;
-import com.sprint.mission.discodeit.dto.UserResponse;
+import com.sprint.mission.discodeit.dto2.UserLoginRequest;
+import com.sprint.mission.discodeit.dto2.UserResponse;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.AuthService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class BasicAuthService implements AuthService {
     private final UserRepository userRepository;
-
-    public BasicAuthService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
 
     @Override
     public UserResponse login(UserLoginRequest request) {
